DEV: initial support for prettier-ruby

+++ app/models/optimized_image.rb
@@ -1,7 +1,7 @@
-require_dependency "file_helper"
-require_dependency "url_helper"
-require_dependency "db_helper"
-require_dependency "file_store/local_store"
+require_dependency 'file_helper'
+require_dependency 'url_helper'
+require_dependency 'db_helper'
+require_dependency 'file_store/local_store'
 
 class OptimizedImage < ActiveRecord::Base
   belongs_to :upload
@@ -10,20 +10,25 @@ class OptimizedImage < ActiveRecord::Base
   VERSION = 2
 
   def self.lock(upload_id, width, height)
-    @hostname ||= `hostname`.strip rescue "unknown"
+    @hostname ||=
+      begin
+        `hostname`.strip
+      rescue StandardError
+        'unknown'
+      end
     # note, the extra lock here ensures we only optimize one image per machine on webs
     # this can very easily lead to runaway CPU so slowing it down is beneficial and it is hijacked
     #
     # we can not afford this blocking in Sidekiq cause it can lead to starvation
     if Sidekiq.server?
-      DistributedMutex.synchronize("optimized_image_#{upload_id}_#{width}_#{height}") do
-        yield
-      end
+      DistributedMutex.synchronize(
+        "optimized_image_#{upload_id}_#{width}_#{height}"
+      ) { yield }
     else
       DistributedMutex.synchronize("optimized_image_host_#{@hostname}") do
-        DistributedMutex.synchronize("optimized_image_#{upload_id}_#{width}_#{height}") do
-          yield
-        end
+        DistributedMutex.synchronize(
+          "optimized_image_#{upload_id}_#{width}_#{height}"
+        ) { yield }
       end
     end
   end
@@ -33,11 +38,9 @@ class OptimizedImage < ActiveRecord::Base
     return if upload.try(:sha1).blank?
 
     # no extension so try to guess it
-    if (!upload.extension)
-      upload.fix_image_extension
-    end
+    upload.fix_image_extension if (!upload.extension)
 
-    if !upload.extension.match?(IM_DECODERS) && upload.extension != "svg"
+    if !upload.extension.match?(IM_DECODERS) && upload.extension != 'svg'
       if !opts[:raise_on_error]
         # nothing to do ... bad extension, not an image
         return
@@ -67,24 +70,30 @@ class OptimizedImage < ActiveRecord::Base
       # create the thumbnail otherwise
       original_path = Discourse.store.path_for(upload)
       if original_path.blank?
-        external_copy = Discourse.store.download(upload) rescue nil
+        external_copy =
+          begin
+            Discourse.store.download(upload)
+          rescue StandardError
+            nil
+          end
         original_path = external_copy.try(:path)
       end
 
       if original_path.blank?
-        Rails.logger.error("Could not find file in the store located at url: #{upload.url}")
-      else
+        Rails.logger.error(
+          "Could not find file in the store located at url: #{upload.url}"
+        )
+
         # create a temp file with the same extension as the original
+      else
         extension = ".#{opts[:format] || upload.extension}"
 
-        if extension.length == 1
-          return nil
-        end
+        return nil if extension.length == 1
 
-        temp_file = Tempfile.new(["discourse-thumbnail", extension])
+        temp_file = Tempfile.new(['discourse-thumbnail', extension])
         temp_path = temp_file.path
 
-        if upload.extension == "svg"
+        if upload.extension == 'svg'
           FileUtils.cp(original_path, temp_path)
           resized = true
         elsif opts[:crop]
@@ -94,16 +103,17 @@ class OptimizedImage < ActiveRecord::Base
         end
 
         if resized
-          thumbnail = OptimizedImage.create!(
-            upload_id: upload.id,
-            sha1: Upload.generate_digest(temp_path),
-            extension: extension,
-            width: width,
-            height: height,
-            url: "",
-            filesize: File.size(temp_path),
-            version: VERSION
-          )
+          thumbnail =
+            OptimizedImage.create!(
+              upload_id: upload.id,
+              sha1: Upload.generate_digest(temp_path),
+              extension: extension,
+              width: width,
+              height: height,
+              url: '',
+              filesize: File.size(temp_path),
+              version: VERSION
+            )
 
           # store the optimized image and update its url
           File.open(temp_path) do |file|
@@ -120,9 +130,7 @@ class OptimizedImage < ActiveRecord::Base
       end
 
       # make sure we remove the cached copy from external stores
-      if Discourse.store.external?
-        external_copy&.close
-      end
+      external_copy&.close if Discourse.store.external?
 
       thumbnail
     end
@@ -136,7 +144,7 @@ class OptimizedImage < ActiveRecord::Base
   end
 
   def local?
-    !(url =~ /^(https?:)?\/\//)
+    !(url =~ %r{^(https?:)?\/\/})
   end
 
   def calculate_filesize
@@ -155,12 +163,15 @@ class OptimizedImage < ActiveRecord::Base
     else
       # we may have a bad optimized image so just skip for now
       # and do not break here
-      size = calculate_filesize rescue nil
+      size =
+        begin
+          calculate_filesize
+        rescue StandardError
+          nil
+        end
 
       write_attribute(:filesize, size)
-      if !new_record?
-        update_columns(filesize: size)
-      end
+      update_columns(filesize: size) if !new_record?
       size
     end
   end
@@ -169,14 +180,12 @@ class OptimizedImage < ActiveRecord::Base
     # this matches instructions which call #to_s
     path = path.to_s
     return false if path != File.expand_path(path)
-    return false if path !~ /\A[\w\-\.\/]+\z/m
+    return false if path !~ %r{\A[\w\-\.\/]+\z}m
     true
   end
 
   def self.ensure_safe_paths!(*paths)
-    paths.each do |path|
-      raise Discourse::InvalidAccess unless safe_path?(path)
-    end
+    paths.each { |path| raise Discourse::InvalidAccess unless safe_path?(path) }
   end
 
   IM_DECODERS ||= /\A(jpe?g|png|ico|gif)\z/i
@@ -196,12 +205,14 @@ class OptimizedImage < ActiveRecord::Base
       extension = File.extname(opts[:filename] || ext_path || path)[1..-1]
     end
 
-    raise Discourse::InvalidAccess if !extension || !extension.match?(IM_DECODERS)
+    if !extension || !extension.match?(IM_DECODERS)
+      raise Discourse::InvalidAccess
+    end
     "#{extension}:#{path}"
   end
 
   def self.thumbnail_or_resize
-    SiteSetting.strip_image_metadata ? "thumbnail" : "resize"
+    SiteSetting.strip_image_metadata ? 'thumbnail' : 'resize'
   end
 
   def self.resize_instructions(from, to, dimensions, opts = {})
@@ -213,37 +224,48 @@ class OptimizedImage < ActiveRecord::Base
 
     instructions = ['convert', "#{from}[0]"]
 
-    if opts[:colors]
-      instructions << "-colors" << opts[:colors].to_s
-    end
+    instructions << '-colors' << opts[:colors].to_s if opts[:colors]
 
     # NOTE: ORDER is important!
-    instructions.concat(%W{
-      -auto-orient
-      -gravity center
-      -background transparent
-      -#{thumbnail_or_resize} #{dimensions}^
-      -extent #{dimensions}
-      -interpolate catrom
-      -unsharp 2x0.5+0.7+0
-      -interlace none
-      -quality 98
-      -profile #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
-      #{to}
-    })
+    instructions.concat(
+      %W[
+        -auto-orient
+        -gravity
+        center
+        -background
+        transparent
+        -#{thumbnail_or_resize}
+        #{dimensions}^
+        -extent
+        #{dimensions}
+        -interpolate
+        catrom
+        -unsharp
+        2x0.5+0.7+0
+        -interlace
+        none
+        -quality
+        98
+        -profile
+        #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
+        #{to}
+      ]
+    )
   end
 
   def self.resize_instructions_animated(from, to, dimensions, opts = {})
     ensure_safe_paths!(from, to)
 
-    %W{
+    %W[
       gifsicle
       --colors=#{opts[:colors] || 256}
-      --resize-fit #{dimensions}
+      --resize-fit
+      #{dimensions}
       --optimize=3
-      --output #{to}
+      --output
+      #{to}
       #{from}
-    }
+    ]
   end
 
   def self.crop_instructions(from, to, dimensions, opts = {})
@@ -252,33 +274,43 @@ class OptimizedImage < ActiveRecord::Base
     from = prepend_decoder!(from, to, opts)
     to = prepend_decoder!(to, to, opts)
 
-    %W{
+    %W[
       convert
       #{from}[0]
       -auto-orient
-      -gravity north
-      -background transparent
-      -#{thumbnail_or_resize} #{opts[:width]}
-      -crop #{dimensions}+0+0
-      -unsharp 2x0.5+0.7+0
-      -interlace none
-      -quality 98
-      -profile #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
+      -gravity
+      north
+      -background
+      transparent
+      -#{thumbnail_or_resize}
+      #{opts[:width]}
+      -crop
+      #{dimensions}+0+0
+      -unsharp
+      2x0.5+0.7+0
+      -interlace
+      none
+      -quality
+      98
+      -profile
+      #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
       #{to}
-    }
+    ]
   end
 
   def self.crop_instructions_animated(from, to, dimensions, opts = {})
     ensure_safe_paths!(from, to)
 
-    %W{
+    %W[
       gifsicle
-      --crop 0,0+#{dimensions}
+      --crop
+      0,0+#{dimensions}
       --colors=#{opts[:colors] || 256}
       --optimize=3
-      --output #{to}
+      --output
+      #{to}
       #{from}
-    }
+    ]
   end
 
   def self.downsize_instructions(from, to, dimensions, opts = {})
@@ -287,17 +319,22 @@ class OptimizedImage < ActiveRecord::Base
     from = prepend_decoder!(from, to, opts)
     to = prepend_decoder!(to, to, opts)
 
-    %W{
+    %W[
       convert
       #{from}[0]
       -auto-orient
-      -gravity center
-      -background transparent
-      -interlace none
-      -resize #{dimensions}
-      -profile #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
+      -gravity
+      center
+      -background
+      transparent
+      -interlace
+      none
+      -resize
+      #{dimensions}
+      -profile
+      #{File.join(Rails.root, 'vendor', 'data', 'RT_sRGB.icm')}
       #{to}
-    }
+    ]
   end
 
   def self.downsize_instructions_animated(from, to, dimensions, opts = {})
@@ -305,22 +342,22 @@ class OptimizedImage < ActiveRecord::Base
   end
 
   def self.resize(from, to, width, height, opts = {})
-    optimize("resize", from, to, "#{width}x#{height}", opts)
+    optimize('resize', from, to, "#{width}x#{height}", opts)
   end
 
   def self.crop(from, to, width, height, opts = {})
     opts[:width] = width
-    optimize("crop", from, to, "#{width}x#{height}", opts)
+    optimize('crop', from, to, "#{width}x#{height}", opts)
   end
 
   def self.downsize(from, to, dimensions, opts = {})
-    optimize("downsize", from, to, dimensions, opts)
+    optimize('downsize', from, to, dimensions, opts)
   end
 
   def self.optimize(operation, from, to, dimensions, opts = {})
     method_name = "#{operation}_instructions"
     if !!opts[:allow_animation] && (from =~ /\.GIF$/i)
-      method_name += "_animated"
+      method_name += '_animated'
     end
     instructions = self.send(method_name.to_sym, from, to, dimensions, opts)
     convert_with(instructions, to, opts)
@@ -329,22 +366,19 @@ class OptimizedImage < ActiveRecord::Base
   MAX_PNGQUANT_SIZE = 500_000
 
   def self.convert_with(instructions, to, opts = {})
-    Discourse::Utils.execute_command("nice", "-n", "10", *instructions)
+    Discourse::Utils.execute_command('nice', '-n', '10', *instructions)
 
-    allow_pngquant = to.downcase.ends_with?(".png") && File.size(to) < MAX_PNGQUANT_SIZE
+    allow_pngquant =
+      to.downcase.ends_with?('.png') && File.size(to) < MAX_PNGQUANT_SIZE
     FileHelper.optimize_image!(to, allow_pngquant: allow_pngquant)
     true
   rescue => e
     if opts[:raise_on_error]
       raise e
     else
-      error = +"Failed to optimize image:"
+      error = +'Failed to optimize image:'
 
-      if e.message =~ /^convert:([^`]+)/
-        error << $1
-      else
-        error << " unknown reason"
-      end
+      e.message =~ /^convert:([^`]+)/ ? error << $1 : error << ' unknown reason'
 
       Discourse.warn(error, location: to, error_message: e.message)
       false
@@ -358,27 +392,33 @@ class OptimizedImage < ActiveRecord::Base
       max_file_size_kb = SiteSetting.max_image_size_kb.kilobytes
       local_store = FileStore::LocalStore.new
 
-      scope = OptimizedImage.includes(:upload)
-        .where("url NOT LIKE '%/optimized/_X/%'")
-        .order(id: :desc)
+      scope =
+        OptimizedImage.includes(:upload).where(
+          "url NOT LIKE '%/optimized/_X/%'"
+        )
+          .order(id: :desc)
 
       scope.limit(limit) if limit
 
       scope.each do |optimized_image|
         begin
-          # keep track of the url
-          previous_url = optimized_image.url.dup
-          # where is the file currently stored?
-          external = previous_url =~ /^\/\//
+          previous_url = optimized_image.url.dup # keep track of the url
+          external = previous_url =~ %r{^\/\/} # where is the file currently stored?
+
           # download if external
           if external
-            url = SiteSetting.scheme + ":" + previous_url
-            file = FileHelper.download(
-              url,
-              max_file_size: max_file_size_kb,
-              tmp_file_name: "discourse",
-              follow_redirect: true
-            ) rescue nil
+            url = SiteSetting.scheme + ':' + previous_url
+            file =
+              begin
+                FileHelper.download(
+                  url,
+                  max_file_size: max_file_size_kb,
+                  tmp_file_name: 'discourse',
+                  follow_redirect: true
+                )
+              rescue StandardError
+                nil
+              end
             path = file.path
           else
             path = local_store.path_for(optimized_image)
@@ -392,16 +432,20 @@ class OptimizedImage < ActiveRecord::Base
           FileHelper.optimize_image!(path)
           # store to new location & update the filesize
           File.open(path) do |f|
-            optimized_image.url = Discourse.store.store_optimized_image(f, optimized_image)
+            optimized_image.url =
+              Discourse.store.store_optimized_image(f, optimized_image)
             optimized_image.save
           end
           # remap the URLs
-          DbHelper.remap(UrlHelper.absolute(previous_url), optimized_image.url) unless external
+          unless external
+            DbHelper.remap(
+              UrlHelper.absolute(previous_url),
+              optimized_image.url
+            )
+          end
           DbHelper.remap(previous_url, optimized_image.url)
           # remove the old file (when local)
-          unless external
-            FileUtils.rm(path, force: true)
-          end
+          FileUtils.rm(path, force: true) unless external
         rescue => e
           problems << { optimized_image: optimized_image, ex: e }
           # just ditch the optimized image if there was any errors
@@ -415,7 +459,6 @@ class OptimizedImage < ActiveRecord::Base
 
     problems
   end
-
 end
 
 # == Schema Information
