+++ db/fixtures/002_groups.rb
@@ -1,6 +1,6 @@
 Group.ensure_automatic_groups!
-if g = Group.find_by(name: 'trust_level_5', id: 15)
-  g.destroy!
-end
+g.destroy! if g = Group.find_by(name: 'trust_level_5', id: 15)
 
-Group.where(name: 'everyone').update_all(visibility_level: Group.visibility_levels[:owners])
+Group.where(name: 'everyone').update_all(
+  visibility_level: Group.visibility_levels[:owners]
+)
